use off_t for seek and tell operations

+++ oslib/osfile.h
@@ -19,12 +19,12 @@ void osClose(osFile os);
 int osGetChar(osFile os);
 uint32_t osRead(osFile os,void *buf,uint32_t bytes);
 uint32_t osFGets(osFile os,char *str,uint32_t max);
-uint32_t osFTell(osFile os);
+off_t osFTell(osFile os);
 bool osPutChar(osFile os, char ch);
 bool osWrite(osFile os,const void *buf, uint32_t bytes);
 bool osPuts(osFile os,char *str);
 bool osFPrintf(osFile os,char *fmt,...);
 bool osVFPrintf(osFile os,char *fmt,va_list args);
-void osSeek(osFile os,uint32_t offset,short mode);
+void osSeek(osFile os,off_t offset,short mode);
 
 #endif
